restructure website and add layouts
