+++ apps/website/src/pages/_app.tsx
@@ -4,6 +4,7 @@ import '@/styles/reset.css'
 import { Provider } from '@status-im/components'
 import { Inter } from 'next/font/google'
 
+import type { Page, PageLayout } from 'next'
 import type { AppProps } from 'next/app'
 
 const inter = Inter({
@@ -12,12 +13,16 @@ const inter = Inter({
   subsets: ['latin'],
 })
 
-export default function App({ Component, pageProps }: AppProps) {
+type Props = AppProps & {
+  Component: Page
+}
+
+export default function App({ Component, pageProps }: Props) {
+  const getLayout: PageLayout = Component.getLayout || (page => page)
+
   return (
     <div id="app" className={inter.variable}>
-      <Provider>
-        <Component {...pageProps} />
-      </Provider>
+      <Provider>{getLayout(<Component {...pageProps} />)}</Provider>
     </div>
   )
 }
