+++ apps/website/tailwind.config.js
@@ -4,13 +4,8 @@ const colors = require('@status-im/colors')
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: [
+    './src/**/*.{js,ts,jsx,tsx,mdx}',
     // './app/**/*.{js,ts,jsx,tsx,mdx}',
-    './pages/**/*.{js,ts,jsx,tsx,mdx}',
-    './components/**/*.{js,ts,jsx,tsx,mdx}',
-
-    // todo: move all to `src` directory
-    // // Or if using `src` directory:
-    // './src/**/*.{js,ts,jsx,tsx,mdx}',
   ],
   theme: {
     fontFamily: {
