+++ apps/website/tsconfig.json
@@ -10,7 +10,7 @@
     "baseUrl": ".",
     "paths": {
       "react-native": ["react-native-web"],
-      "@/*": ["./*"]
+      "@/*": ["./src/*"]
       // "@status-im/*": ["./node_modules/@status-im/*"]
       // "@status-im/js": ["./node_modules/@status-im/js/packages/status-js"],
       // "@status-im/components": [
